backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai==1.30.1
faiss-cpu==1.8.0
pypdf2==3.0.1
tiktoken==0.7.0
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
aiofiles==23.2.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
